Added docker volume mounting (#277)
* Added docker volume mounting

* Added tests, command attribute

* yapf'ed

* Fixed instance where molecule always try to check for vagrant instances

* Fixed bug where molecule overwrote custom config file. Moved instance name update to vagrant

* put populate_instance_names in vagrantprovisioner. fixed --platform all to actually use different boxes for each instance

* Fixed tests.

* Remove extraneous #

* Added check for vagrant executable in vagrant tests

* Attempted to move template_creation and then reverted back

* removed unused imports

* added customconf state test

* Fixed formatting

+++ molecule/core.py
@@ -48,6 +48,7 @@ class Molecule(object):
         self._provisioner = None
 
     def main(self):
+
         if not os.path.exists(self._config.config['molecule']['molecule_dir']):
             os.makedirs(self._config.config['molecule']['molecule_dir'])
 
@@ -181,7 +182,8 @@ class Molecule(object):
         :return: None
         """
         os.remove(self._config.config['molecule']['rakefile_file'])
-        os.remove(self._config.config['ansible']['config_file'])
+        if self._state.customconf is False:
+            os.remove(self._config.config['ansible']['config_file'])
 
     def _create_templates(self):
         """
@@ -192,9 +194,13 @@ class Molecule(object):
         # ansible.cfg
         kwargs = {'molecule_dir':
                   self._config.config['molecule']['molecule_dir']}
-        utilities.write_template(
-            self._config.config['molecule']['ansible_config_template'],
-            self._config.config['ansible']['config_file'], kwargs=kwargs)
+        if not os.path.isfile(self._config.config['ansible']['config_file']):
+            utilities.write_template(
+                self._config.config['molecule']['ansible_config_template'],
+                self._config.config['ansible']['config_file'], kwargs=kwargs)
+            self._state.change_state('customconf', False)
+        else:
+            self._state.change_state('customconf', True)
 
         # rakefile
         kwargs = {
